Fixed mlx and map-parsing leaks, all blocks being freed
